requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
Django==3.2.3
djangorestframework==3.12.4
Flask==2.0.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pyTelegramBotAPI==3.7.9
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
Werkzeug==2.0.1
